Use signalfd() if available
signalfd() is designed to do what the pipe "trick" does for handling
signals asynchronously, so use signalfd() if available (since Linux
2.6.26).

+++ keepalived/vrrp/vrrp_scheduler.c
@@ -942,8 +942,8 @@ vrrp_script_child_thread(thread_t * thread)
 	wait_status = THREAD_CHILD_STATUS(thread);
 
 	if (WIFEXITED(wait_status)) {
-		int status;
-		status = WEXITSTATUS(wait_status);
+		int8_t status;
+		status = (int8_t)WEXITSTATUS(wait_status);
 
 		/* Report if status has changed */
 		if (status != vscript->last_status) {
